% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{add.cluster}
\alias{add.cluster}
\title{Add an empty cluster to a metadata object}
\usage{
add.cluster(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A metadata object with an empty additional cluster
}
\description{
Add an empty cluster to a metadata object
}
\examples{
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5), Sigma=diag(1,2)),
                         c2 = list(n = 25, mu = c(-1,-2), Sigma=diag(1,2))),
         dist = mvrnorm)
m2 <- add.cluster(m)
}

