% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{addCluster}
\alias{addCluster}
\title{Add an empty cluster to a metadata object}
\usage{
addCluster(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A metadata object with an empty additional cluster
}
\description{
Add an empty cluster to a metadata object
}
\examples{
require(MASS)
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5), Sigma=diag(1,2)),
                         c2 = list(n = 25, mu = c(-1,-2), Sigma=diag(1,2))),
         genfunc = mvrnorm)
m2 <- addCluster(m)
}

