% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkSetup}
\alias{checkSetup}
\title{Performs various consistency checks on a setup file}
\usage{
checkSetup(file)
}
\arguments{
\item{file}{A .R file with a new simulation setup}
}
\description{
Performs various consistency checks on a setup file
}

