% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{createFileskeleton}
\alias{createFileskeleton}
\title{Create a new setup file template}
\usage{
createFileskeleton(newname, mail, inst, author, type = c("metric",
  "functional", "ordinal", "binary", "randomstring", "wordnet"),
  infotable = NULL, ref = "Unpublished", codefile = F)
}
\arguments{
\item{newname}{The name of the new setup (and subsequently the file name)}

\item{mail}{The contact e-mail address of the author}

\item{inst}{The institution of the author}

\item{author}{The full name of the author}

\item{type}{The data type of this setup}

\item{infotable}{The setup summary table}

\item{ref}{The reference to the publication where the setup was used}

\item{codefile}{If functions that are needed for the data generation of the setup are stored in some other .R file, the path can be supplied}
}
\description{
Create a new setup file template
}

