% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{delete.cluster}
\alias{delete.cluster}
\title{Delete a cluster from a metadata object}
\usage{
delete.cluster(m, clnumber)
}
\arguments{
\item{m}{A metadata object}

\item{clnumber}{The cluster to delete}
}
\value{
A metadata object
}
\description{
Delete a cluster from a metadata object
}
\examples{
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5), Sigma=diag(1,2)),
                         c2 = list(n = 25, mu = c(-1,-2), Sigma=diag(1,2))),
         dist = mvrnorm)
m2 <- delete.cluster(m, 2)
}

