% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate.data}
\alias{generate.data}
\title{Generate a dataset from a metadata object}
\usage{
generate.data(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A dataset as specified by the metadata object
}
\description{
Generate a dataset from a metadata object
}
\examples{
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5), Sigma=diag(1,2)),
                         c2 = list(n = 25, mu = c(-1,-2), Sigma=diag(1,2))),
         dist = mvrnorm)
generate.data(m)
}

