% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{methods}
\name{generateData,metadata.functional-method}
\alias{generateData,metadata.functional-method}
\title{Generate a dataset from a metadata object}
\usage{
\S4method{generateData}{metadata.functional}(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A dataset as specified by the metadata object
}
\description{
Generate a dataset from a metadata object
}

