% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{initializeObject}
\alias{initializeObject}
\title{Initialize a new metadata object}
\usage{
initializeObject(type, k, genfunc, seed = list(100, paste(R.version$major,
  R.version$minor, sep = "."), RNGkind()))
}
\arguments{
\item{type}{The data type for the new object}

\item{k}{Number of clusters}

\item{genfunc}{The distribution function for data generation}

\item{seed}{The random number seed parameters for the data generation}
}
\value{
A metadata object
}
\description{
Initialize a new metadata object
}
\examples{
require(MASS)
initializeObject(type = "metric", k = 3, genfunc = mvrnorm)
}

