% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.binary-class}
\alias{metadata.binary-class}
\title{A class that represents a metadata object for binary data}
\description{
A class that represents a metadata object for binary data
}

