% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.functional-class}
\alias{metadata.functional-class}
\title{A class that represents a metadata object for functional data}
\description{
A class that represents a metadata object for functional data
}

