% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.general-class}
\alias{metadata.general-class}
\title{A class to represent a metadata object}
\description{
A class to represent a metadata object
}
\section{Fields}{

\describe{
\item{\code{clusters}}{A list of cluster information}

\item{\code{dist}}{A string specifying a distribution for the random numbers}

\item{\code{seedinfo}}{A list with the parameters for the random number generator}
}}

