% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.metric-class}
\alias{metadata.metric-class}
\title{A class that represents a metadata object for metric data}
\description{
A class that represents a metadata object for metric data
}
\section{Fields}{

\describe{
\item{\code{standardization}}{If standardization is needed, function can be supplied}
}}

