% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.ordinal-class}
\alias{metadata.ordinal-class}
\title{A class that represents a metadata object for ordinal data}
\description{
A class that represents a metadata object for ordinal data
}

