% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metadata.randomstring-class}
\alias{metadata.randomstring-class}
\title{A class that represents a metadata object for string data}
\description{
A class that represents a metadata object for string data
}

