% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{metaplot,metadata.ordinal-method}
\alias{metaplot,metadata.ordinal-method}
\title{Plot a metadata object}
\usage{
\S4method{metaplot}{metadata.ordinal}(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A plot, created by generating an instance of the dataset from the metadata object
}
\description{
Plot a metadata object
}

