% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{plot3dMetadata,metadata.metric-method}
\alias{plot3dMetadata,metadata.metric-method}
\title{3d plot of a metric metadata object}
\usage{
\S4method{plot3dMetadata}{metadata.metric}(m)
}
\arguments{
\item{m}{A metadata object (for metric data)}
}
\value{
A 3d plot using function \code{plot3d} from package \code{rgl}
}
\description{
3d plot of a metric metadata object
}

