% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot3dMetadata}
\alias{plot3dMetadata}
\title{3d plot of a metric metadata object}
\usage{
plot3dMetadata(m)
}
\arguments{
\item{m}{A metadata object (for metric data)}
}
\value{
A 3d plot using function \code{plot3d} from package \code{rgl}
}
\description{
3d plot of a metric metadata object
}
\examples{
require(MASS)
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5,4), Sigma=diag(1,3)),
                         c2 = list(n = 25, mu = c(-1,-2,-2), Sigma=diag(1,3))),
         genfunc = mvrnorm)
plot3dMetadata(m)
}

