% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotMetadata,metadata.metric-method}
\alias{plotMetadata,metadata.metric-method}
\title{Plot a metadata object}
\usage{
\S4method{plotMetadata}{metadata.metric}(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A plot, created by generating an instance of the dataset from the metadata object
}
\description{
Plot a metadata object
}
