% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotMetadata}
\alias{plotMetadata}
\title{Plot a metadata object}
\usage{
plotMetadata(m)
}
\arguments{
\item{m}{A metadata object}
}
\value{
A plot, created by generating an instance of the dataset from the metadata object
}
\description{
Plot a metadata object
}
\examples{
require(MASS)
m <- new("metadata.metric", 
         clusters = list(c1 = list(n = 25, mu = c(4,5), Sigma=diag(1,2)),
                         c2 = list(n = 25, mu = c(-1,-2), Sigma=diag(1,2))),
         genfunc = mvrnorm)
plotMetadata(m)
}

