% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sampleGrid}
\alias{sampleGrid}
\title{Sample grid points for functional data}
\usage{
sampleGrid(total_n, minT, maxT, granularity, regular = FALSE)
}
\arguments{
\item{total_n}{Number of Observations}

\item{minT}{Minimum number of time points sampled}

\item{maxT}{Maximum number of time points sampled}

\item{granularity}{Number of possible time points in total}

\item{regular}{If TRUE, maxT time points are sampled at the same time points for each function}
}
\value{
A binary matrix indicating whether the function should be evaluated at a given time point
}
\description{
Sample grid points for functional data
}
\examples{
sampleGrid(total_n = 10, minT = 4, maxT = 10, granularity = 20)
}

