% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setupsummary}
\alias{setupsummary}
\title{Returns the setup summary}
\usage{
setupsummary(name)
}
\arguments{
\item{name}{The name of the setup}
}
\value{
The summary table of \code{name}
}
\description{
Returns the setup summary
}
\examples{
setupsummary(dangl2014)
}

