% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summarizeSetup}
\alias{summarizeSetup}
\title{Returns the setup summary}
\usage{
summarizeSetup(name)
}
\arguments{
\item{name}{The name of the setup}
}
\value{
The summary table of \code{name}
}
\description{
Returns the setup summary
}

